templates/requirements.txt
Flask
opencv-python
numpy
scikit-image
torch
diffusers
transformers
accelerate
Pillow